render_requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0 